pyproject.toml
[project]
name = "nerc_rates"
authors = [
  {name="MOC Alliance"},
]
version = "0.1"
readme = "README.md"
classifiers = [
  "Programming Language :: Python :: 3"
]
requires-python = ">=3.8"
dependencies = [
  "pydantic",
  "pyyaml",
  "requests",
]

[build-system]
requires = [
  "setuptools>=42",
  "wheel"
]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-dir]
"" = "src"
